backend/requirements.txt
google-generativeai==0.8.3
google-genai==1.41.0
pillow>=10.3.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
playwright>=1.46.0
google-search-results==2.4.2
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pydantic>=2.8.0
